% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extract_function_code}
\alias{extract_function_code}
\title{Internal utility function}
\usage{
extract_function_code(input_string)
}
\description{
This function is for internal package use only
}
\keyword{internal}
