% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extract_function_info}
\alias{extract_function_info}
\title{Internal utility function}
\usage{
extract_function_info(file_path)
}
\description{
This function is for internal package use only
}
\keyword{internal}
