% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{generate_llm_prompt}
\alias{generate_llm_prompt}
\title{Internal utility function}
\usage{
generate_llm_prompt(functions_info)
}
\description{
This function is for internal package use only
}
\keyword{internal}
