% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSorcery-package.R
\docType{package}
\name{RSorcery-package}
\alias{RSorcery}
\alias{RSorcery-package}
\title{RSorcery: Magically generates functions based on decorators to function definitions}
\description{
Package generate functions automatically from comments made on top of functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kriskumar/RSorcery}
  \item Report bugs at \url{https://github.com/kriskumar/RSorcery/issues}
}

}
\author{
\strong{Maintainer}: Krishna Kumar \email{kk2250+rsorcery@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
