% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dosorcery}
\alias{dosorcery}
\title{Takes a file with a list of function definitions which are preceded
by a comment line with what the function does.
This function is exported and available to users.}
\usage{
dosorcery(file_path, output_file = NULL)
}
\description{
Takes a file with a list of function definitions which are preceded
by a comment line with what the function does.
This function is exported and available to users.
}
